dev local: make clearer by moving into a custom hook
